Crate urlshortener

urlshortener

An easy library for retrieving short urls.

Installation

Add the following dependency to your project's Cargo.toml:

[dependencies]
urlshortener = "0.6"

And add this to your root file:

extern crate urlshortener;Run

Usage

Creating a short URL via a specified provider is very simple:

use urlshortener::{Provider, UrlShortener};

let us = UrlShortener::new();
let short_url = us.generate("https://my-long-url.com", Provider::IsGd);
assert!(short_url.is_ok());Run

Or attempting all URL shorteners until one is successfully generated:

use urlshortener::UrlShortener;

let us = UrlShortener::new();
let short_url = us.try_generate("https://my-long-url.com");
assert!(short_url.is_ok());Run

In order to use service with authentication use the appropriate provider directly:

use urlshortener::{ UrlShortener, Provider };

let us = UrlShortener::new();
let key = "MY_API_KEY";
let short_url = us.generate("https://my-long-url.com", Provider::GooGl { api_key:
key.to_owned() });
assert!(short_url.is_ok());Run

Structs

UrlShortener
Url shortener: the way to retrieve a short url.

Enums

Provider
Used to specify which provider to use to generate a short URL.

Functions

providers
Returns a vector of all Provider variants which do not require authentication. This list is in order of provider quality.