Module urlshortener::providers

Library service providers implementation.

Enums

Provider
Used to specify which provider to use to generate a short URL.

Functions

parse
Parses the response from a successful request to a provider into the URL-shortened string.

providers
Returns a vector of all Provider variants. This list is in order of provider quality.

request
Performs a request to the short link provider. Response to be parsed or None on a error.