Struct urlshortener::UrlShortener

pub struct UrlShortener {
    // some fields omitted
}

Url shortener: the way to retrieve a short url.

Methods

impl UrlShortener

fn new() -> UrlShortener

Creates new UrlShortener with default (3 seconds) timeout.

fn with_timeout(seconds: u64) -> UrlShortener

Creates new UrlShortener with custom read timeout.

fn try_generate<S: Into<String>>(&self, url: S) -> Result<String, Error>

Try to generate a short URL from each provider, iterating over each provider until a short URL is successfully generated.

Examples

use urlshortener::UrlShortener;

let us = UrlShortener::new();
let long_url = "https://rust-lang.org";
let _short_url = us.try_generate(long_url);

Errors

Returns an Error<ErrorKind::Other> if there is an error generating a short URL from all providers.

fn generate<S: Into<String>>(&self, url: S, provider: Provider) -> Result<String, Error>

Attempts to get a short URL using the specified provider.

Examples

use urlshortener::{Provider, UrlShortener};

let us = UrlShortener::new();
let long_url = "http://rust-lang.org";
let _short_url = us.generate(long_url, Provider::IsGd);

Errors

Returns an std::io::Error if there is an error generating a short URL from the given provider due to either:

a. a decode error (ErrorKind::Other); b. the service being unavailable (ErrorKind::ConnectionAborted)