Enum urlshortener::Provider

pub enum Provider {
    Abv8,
    BamBz,
    Bmeo,
    BnGy,
    FifoCc,
    HecSu,
    HmmRs,
    IsGd,
    NowLinks,
    PhxCoIn,
    PsbeCo,
    SCoop,
    Rdd,
    Rlu,
    SirBz,
    TinyUrl,
    TinyPh,
    TnyIm,
    UrlShortenerIo,
    VGd,
}

Used to specify which provider to use to generate a short URL.

Variants
Abv8

http://abv8.me provider

Notes:

You may not shorten more than 20 unique URLs within a 3-minute period.

You may not shorten more than 60 unique URLs within a 15-minute period.

BamBz

https://bam.bz provider

Bmeo

http://bmeo.org provider

BnGy

https://bn.gy provider

FifoCc

http://fifo.cc provider

HecSu

https://hec.su provider

Notes:

Limited to 3000 API requests per day

HmmRs

http://hmm.rs provider

IsGd

https://is.gd provider

NowLinks

http://nowlinks.net provider

PhxCoIn

http://phx.co.in provider

Notes:

After some time the service will display ads

Instead of redirecting, a preview page will be displayed

Currently unstable

PsbeCo

http://psbe.co provider

SCoop

http://s.coop provider

Rdd

http://readbility.com provider

Rlu

http://rlu.ru provider

Notes:

If you send a lot of requests from one IP, it can be blocked. If you plan to add more then 100 URLs in one hour, please let the technical support know. Otherwise your IP can be blocked unexpectedly. Prior added URLs can be deleted.

SirBz

http://sirbz.com provider

Notes:

By default, you are limited to 250 requests per 15 minutes.

TinyUrl

http://tinyurl.com provider

Notes:

This service does not provide any API.

The implementation result depends on the service result web page.

TinyPh

http://tiny.ph provider

TnyIm

http://tny.im provider

UrlShortenerIo

http://url-shortener.io provider

VGd

https://v.gd provider

Methods

impl Provider

fn to_name(&self) -> &str

Converts the Provider variant into its domain name equivilant